Add files via upload
This is the last file that we have about the exponential fitting. Read inside the py file the comments to know more about the methods. However I have to say that basically, this method is not numerically stable, and very often gives that the best approximation is still the linear one with parameters a = 1 and b  =  1.